\name{testing-package}
\alias{testing-package}
\alias{testing}
\docType{package}
\title{\packageTitle{testing}}
\description{\packageDescription{testing}}
\section{Package Content}{\packageIndices{testing}}
\author{\packageAuthor{testing}}
\section{Maintainer}{\packageMaintainer{testing}}
\keyword{package}
